editing copper thickness
I have added updated text in the inventory to specifically state the board should have either the 1oz or 2oz sizes (using all three thickness measures).

I didn't change the text in the attack since I expect that making it clear in the inventory should cover the choices.

I edited the text to bring the one line into proper alignment in the attacks (4 of 5 needed it)

This should close #12