Update 03-xx-Fingerprint_attack.adoc
I removed this once I realized what it was. I had added this originally for the PCB and accidentally copied it over to the remaining attacks. This is only needed for the PCB (it could be left with a N/A, but I think it is better to just remove it)